+++ src/ResultExtractor.hpp
@@ -29,7 +29,7 @@ class ResultExtractor : public Result::Reader::Consumer
 public:
   ResultExtractor(const std::string& directory);
 
-  void on_entry_start(uint32_t entry_number,
+  void on_entry_start(uint8_t entry_number,
                       Result::FileType file_type,
                       uint64_t file_len,
                       nonstd::optional<std::string> raw_file) override;
