+++ src/ResultInspector.hpp
@@ -29,7 +29,7 @@ class ResultInspector : public Result::Reader::Consumer
 public:
   ResultInspector(FILE* stream);
 
-  void on_entry_start(uint32_t entry_number,
+  void on_entry_start(uint8_t entry_number,
                       Result::FileType file_type,
                       uint64_t file_len,
                       nonstd::optional<std::string> raw_file) override;
